# manager/api/api-spec.bundled.yaml
openapi: 3.0.0
info:
  version: 0.0.0
  title: MaEVe CSMS
  description: Internal API to interact with the MaEVe CSMS, external clients should use OCPI.
  contact:
    name: MaEVe team
    email: maeve-team@thoughtworks.com
  license:
    name: Apache 2.0
servers:
  - url: http://localhost:9410/api/v0
    description: The local development server
paths:
...
  /cs/{csId}/auth:
    get:
      summary: Returns the authentication details
      description: |
        Returns the details required by the CSMS gateway to determine how to authenticate
        the charge station
      operationId: lookupChargeStationAuth
      parameters:
        - name: csId
          in: path
          description: The charge station identifier
          schema:
            type: string
            maxLength: 28
      responses:
        '200':
          description: Charge station auth response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChargeStationAuth'
        '404':
          description: Unknown charge station
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
...
  /cs/{csId}/status:
    get:
      summary: Get charge station status
      description: |
        Returns the current status of the specified charge station,
        including connection state, last heartbeat, firmware version, and hardware details.
      operationId: getChargeStationStatus
      parameters:
        - name: csId
          in: path
          description: The charge station identifier
          required: true
          schema:
            type: string
            maxLength: 28
      responses:
        '200':
          description: Charge station status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChargeStationStatusResponse'
        '404':
          description: Unknown charge station
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
  /cs/{csId}/connectors:
    get:
      summary: Get connector statuses
      description: |
        Returns the current status of all connectors on the specified charge station,
        including their availability, error codes, and active transactions.
      operationId: getConnectorStatuses
      parameters:
        - name: csId
          in: path
          description: The charge station identifier
          required: true
          schema:
            type: string
            maxLength: 28
      responses:
        '200':
          description: Connector statuses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ConnectorStatusResponse'
        '404':
          description: Unknown charge station
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
...
  /cs/{csId}/monitoring:
    post:
      summary: Set variable monitoring
      description: |
        Configure variable monitoring on a charge station (OCPP 2.0.1 only).
        The charge station will monitor specified variables and send alerts based on the configured thresholds.
      operationId: setVariableMonitoring
      parameters:
        - name: csId
          in: path
          required: true
          description: The charge station identifier
          schema:
            type: string
            maxLength: 28
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetVariableMonitoringRequest'
      responses:
        '202':
          description: Accepted - monitoring configuration initiated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SetVariableMonitoringResponse'
        '404':
          description: Unknown charge station
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
  /cs/{csId}/monitoring/{monitorId}:
    delete:
      summary: Clear variable monitoring
      description: |
        Remove a specific variable monitor from a charge station (OCPP 2.0.1 only).
      operationId: clearVariableMonitoring
      parameters:
        - name: csId
          in: path
          required: true
          description: The charge station identifier
          schema:
            type: string
            maxLength: 28
        - name: monitorId
          in: path
          required: true
          description: The monitor identifier to clear
          schema:
            type: integer
      responses:
        '202':
          description: Accepted - monitoring clear initiated
        '404':
          description: Unknown charge station or monitor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
  /cs/{csId}/monitoring/base:
    post:
      summary: Set monitoring base
      description: |
        Configure the base level of monitoring on a charge station (OCPP 2.0.1 only).
      operationId: setMonitoringBase
      parameters:
        - name: csId
          in: path
          required: true
          description: The charge station identifier
          schema:
            type: string
            maxLength: 28
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetMonitoringBaseRequest'
      responses:
        '202':
          description: Accepted - monitoring base configuration initiated
        '404':
          description: Unknown charge station
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
  /cs/{csId}/monitoring/level:
    post:
      summary: Set monitoring level
      description: |
        Configure the monitoring severity level on a charge station (OCPP 2.0.1 only).
        Only events with severity at or below this level will be reported.
      operationId: setMonitoringLevel
      parameters:
        - name: csId
          in: path
          required: true
          description: The charge station identifier
          schema:
            type: string
            maxLength: 28
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetMonitoringLevelRequest'
      responses:
        '202':
          description: Accepted - monitoring level configuration initiated
        '404':
          description: Unknown charge station
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
  /cs/{csId}/monitoring/report:
    post:
      summary: Get monitoring report
      description: |
        Request a monitoring report from a charge station (OCPP 2.0.1 only).
        The report will be sent asynchronously via NotifyMonitoringReport notifications.
      operationId: getMonitoringReport
      parameters:
        - name: csId
          in: path
          required: true
          description: The charge station identifier
          schema:
            type: string
            maxLength: 28
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GetMonitoringReportRequest'
      responses:
        '202':
          description: Accepted - monitoring report request initiated
        '404':
          description: Unknown charge station
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
  /cs/{csId}/events:
    get:
      summary: Get charge station events
      description: |
        Retrieve events from a charge station, filtered by type and time range.
        Events are collected from NotifyEvent notifications.
      operationId: getChargeStationEvents
      parameters:
        - name: csId
          in: path
          required: true
          description: The charge station identifier
          schema:
            type: string
            maxLength: 28
        - name: eventType
          in: query
          description: Filter by event type
          schema:
            type: string
        - name: startTime
          in: query
          description: Filter events after this time (ISO8601)
          schema:
            type: string
            format: date-time
        - name: endTime
          in: query
          description: Filter events before this time (ISO8601)
          schema:
            type: string
            format: date-time
        - name: limit
          in: query
          description: Maximum number of events to return
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 100
        - name: offset
          in: query
          description: Number of events to skip
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: Events response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventsResponse'
        '404':
          description: Unknown charge station
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
  /cs/{csId}/reports:
    get:
      summary: Get device reports
      description: |
        Retrieve device reports from a charge station, such as configuration inventory
        or monitoring reports. Reports are collected from NotifyReport and NotifyMonitoringReport notifications.
      operationId: getDeviceReports
      parameters:
        - name: csId
          in: path
          required: true
          description: The charge station identifier
          schema:
            type: string
            maxLength: 28
        - name: reportType
          in: query
          description: Filter by report type
          schema:
            type: string
        - name: requestId
          in: query
          description: Filter by request ID
          schema:
            type: integer
        - name: limit
          in: query
          description: Maximum number of reports to return
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 100
        - name: offset
          in: query
          description: Number of reports to skip
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: Reports response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportsResponse'
        '404':
          description: Unknown charge station
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
components:
  schemas:
    ChargeStationAuth:
      type: object
      description: Connection details for a charge station
      required:
        - securityProfile
      properties:
        securityProfile:
          type: integer
          description: |
            The security profile to use for the charge station: * `0` - unsecured transport with basic auth * `1` - TLS with basic auth * `2` - TLS with client certificate
        base64SHA256Password:
          type: string
          maxLength: 64
          description: The base64 encoded, SHA-256 hash of the charge station password
        invalidUsernameAllowed:
          type: boolean
          description: If set to true then an invalid username will not prevent the charge station connecting
...
    Status:
      type: object
      description: HTTP status
      required:
        - status
      properties:
        status:
          type: string
          description: The status description
        error:
          type: string
          description: The error details
...
    ChargeStationStatusResponse:
      type: object
      description: Current status of a charge station
      required:
        - id
        - connected
      properties:
        id:
          type: string
          description: The charge station identifier
        connected:
          type: boolean
          description: Whether the charge station is currently connected
        lastHeartbeat:
          type: string
          format: date-time
          description: Timestamp of the last heartbeat received
        firmwareVersion:
          type: string
          description: Currently installed firmware version
        model:
          type: string
          description: Charge station model
        vendor:
          type: string
          description: Charge station vendor/manufacturer
        serialNumber:
          type: string
          description: Charge station serial number
    ConnectorStatusResponse:
      type: object
      description: Current status of a connector
      required:
        - connectorId
        - status
        - errorCode
      properties:
        connectorId:
          type: integer
          format: int32
          description: The connector ID
        status:
          type: string
          enum:
            - Available
            - Charging
            - Faulted
            - Finishing
            - Preparing
            - Reserved
            - SuspendedEV
            - SuspendedEVSE
            - Unavailable
          description: Current status of the connector
        errorCode:
          type: string
          enum:
            - NoError
            - ConnectorLockFailure
            - EVCommunicationError
            - GroundFailure
            - HighTemperature
            - InternalError
            - LocalListConflict
            - OtherError
            - OverCurrentFailure
            - OverVoltage
            - PowerMeterFailure
            - PowerSwitchFailure
            - ReaderFailure
            - ResetFailure
            - UnderVoltage
            - WeakSignal
          description: Error code
        info:
          type: string
          description: Additional information about the status
        timestamp:
          type: string
          format: date-time
          description: Timestamp when the status was last updated
        currentTransaction:
          type: string
          description: ID of the current transaction (if any)
    SetVariableMonitoringRequest:
      type: object
      description: Request to set variable monitoring on a charge station
      required:
        - monitoringData
      properties:
        monitoringData:
          type: array
          items:
            $ref: '#/components/schemas/MonitoringData'
    MonitoringData:
      type: object
      required:
        - component
        - variable
        - variableMonitoring
      properties:
        component:
          $ref: '#/components/schemas/Component'
        variable:
          $ref: '#/components/schemas/Variable'
        variableMonitoring:
          type: array
          items:
            $ref: '#/components/schemas/VariableMonitoring'
    Component:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          maxLength: 50
        instance:
          type: string
          maxLength: 50
        evse:
          type: object
          properties:
            id:
              type: integer
            connectorId:
              type: integer
    Variable:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          maxLength: 50
        instance:
          type: string
          maxLength: 50
    VariableMonitoring:
      type: object
      required:
        - value
        - type
        - severity
      properties:
        id:
          type: integer
          description: Monitor ID (optional for new monitors, required for updates)
        transaction:
          type: boolean
          description: Whether monitoring is transaction-specific
        value:
          type: number
          format: float
          description: Threshold value or monitoring parameter
        type:
          type: string
          enum:
            - UpperThreshold
            - LowerThreshold
            - Delta
            - Periodic
            - PeriodicClockAligned
          description: Type of monitoring
        severity:
          type: integer
          minimum: 0
          maximum: 9
          description: Severity level (0 = most critical, 9 = least critical)
    SetVariableMonitoringResponse:
      type: object
      required:
        - results
      properties:
        results:
          type: array
          items:
            type: object
            required:
              - status
            properties:
              id:
                type: integer
                description: Monitor ID assigned by the charge station
              status:
                type: string
                enum:
                  - Accepted
                  - UnknownComponent
                  - UnknownVariable
                  - UnsupportedMonitorType
                  - Rejected
                  - Duplicate
                description: Status of the monitoring request
              statusInfo:
                type: string
                description: Additional status information
    SetMonitoringBaseRequest:
      type: object
      required:
        - monitoringBase
      properties:
        monitoringBase:
          type: string
          enum:
            - All
            - FactoryDefault
            - HardWiredOnly
          description: Base level of monitoring to configure
    SetMonitoringLevelRequest:
      type: object
      required:
        - severity
      properties:
        severity:
          type: integer
          minimum: 0
          maximum: 9
          description: Severity level threshold (0 = most critical, 9 = least critical)
    GetMonitoringReportRequest:
      type: object
      required:
        - requestId
      properties:
        requestId:
          type: integer
          description: Unique request identifier for tracking the report
        componentVariable:
          type: array
          description: Optional list of specific components/variables to report on
          items:
            type: object
            required:
              - component
              - variable
            properties:
              component:
                $ref: '#/components/schemas/Component'
              variable:
                $ref: '#/components/schemas/Variable'
    EventsResponse:
      type: object
      required:
        - events
        - total
        - limit
        - offset
      properties:
        events:
          type: array
          items:
            $ref: '#/components/schemas/ChargeStationEvent'
        total:
          type: integer
          description: Total number of events matching the filter
        limit:
          type: integer
          description: Maximum number of events returned
        offset:
          type: integer
          description: Number of events skipped
    ChargeStationEvent:
      type: object
      required:
        - timestamp
        - eventType
      properties:
        timestamp:
          type: string
          format: date-time
          description: When the event occurred (ISO8601)
        eventType:
          type: string
          description: Type of event (e.g., FirmwareUpdated, Rebooted, SecurityLogWasCleared)
        techCode:
          type: string
          description: Technical error/info code
        techInfo:
          type: string
          description: Additional technical information
        eventData:
          type: string
          description: Additional event details
        componentId:
          type: string
          description: Component that generated the event
        variableId:
          type: string
          description: Variable associated with the event
        severity:
          type: integer
          minimum: 0
          maximum: 9
          description: Event severity (0 = critical, 9 = informational)
    ReportsResponse:
      type: object
      required:
        - reports
        - total
        - limit
        - offset
      properties:
        reports:
          type: array
          items:
            $ref: '#/components/schemas/DeviceReport'
        total:
          type: integer
          description: Total number of reports matching the filter
        limit:
          type: integer
          description: Maximum number of reports returned
        offset:
          type: integer
          description: Number of reports skipped
    DeviceReport:
      type: object
      required:
        - requestId
        - generatedAt
      properties:
        requestId:
          type: integer
          description: Request ID that triggered this report
        generatedAt:
          type: string
          format: date-time
          description: When the report was generated (ISO8601)
        reportType:
          type: string
          description: Type of report (e.g., ConfigurationInventory, MonitoringReport)
        reportData:
          type: array
          description: Report data entries
          items:
            type: object
            properties:
              component:
                $ref: '#/components/schemas/Component'
              variable:
                $ref: '#/components/schemas/Variable'
              variableAttribute:
                type: array
                items:
                  type: object
                  properties:
                    type:
                      type: string
                      description: Attribute type (Actual, Target, MinSet, MaxSet)
                    value:
                      type: string
                      description: Current attribute value
                    mutability:
                      type: string
                      enum:
                        - ReadOnly
                        - WriteOnly
                        - ReadWrite
                    persistent:
                      type: boolean
                    constant:
                      type: boolean